# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Platform Gateway API
  version: 1.0.0
  description: |
    **Platform Gateway**は、エンタープライズ級のマルチテナント管理プラットフォームです。
    Keycloak認証、Stripe Connect決済処理、Cloudflare Edge技術を統合した包括的なAPIを提供します。
    
    ## 主な特徴
    - 🔐 **Keycloak OIDC/JWT認証**: エンタープライズ級認証システム
    - 💳 **Stripe Connect統合**: マルチテナント決済処理
    - 🏢 **マルチテナント対応**: 組織単位でのデータ分離
    - 🛡️ **ロールベースアクセス制御**: 細かい権限管理
    - 📊 **リアルタイム分析**: ダッシュボード・メトリクス
    - 🌐 **Cloudflare Edge**: 高速・グローバル配信
    
    ## 認証方式
    - **JWT Bearer Token**: `Authorization: Bearer <token>`
    - **デモモード**: 開発・テスト用の認証なしアクセス
    
    ## レート制限
    - **一般API**: 100リクエスト/分
    - **認証API**: 20リクエスト/分
    - **決済API**: 50リクエスト/分
  contact:
    name: Platform Gateway API Support
    url: https://github.com/rixy-ynxy/rd-ai-agent-generator
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://ac94bb42.platform-gateway.pages.dev/api
    description: 本番環境
  - url: http://localhost:3000/api
    description: 開発環境
security:
  - bearerAuth: []
  - demoMode: []
paths:
...
  /payment/billing-summary:
    get:
      tags: [決済処理]
      summary: 請求概要・使用状況取得
      description: |
        現在のプラン情報、使用状況、次回請求日などの
        請求概要を取得します。
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
          description: 対象テナントID (権限があれば)
      responses:
        200:
          description: 請求概要取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/BillingSummary'
  /payment/methods:
    get:
      tags: [決済処理]
      summary: 決済方法一覧取得
      description: |
        テナントに登録されている決済方法（クレジットカード等）
        の一覧を取得します。
      parameters:
        - name: tenantId
          in: query
          schema:
            type: string
      responses:
        200:
          description: 決済方法一覧取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaymentMethod'

    post:
      tags: [決済処理]
      summary: 決済方法追加
      description: |
        新しい決済方法を追加します。
        Stripe Payment Methodを使用。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_method_id:
                  type: string
                  description: Stripe Payment Method ID
                is_default:
                  type: boolean
                  description: デフォルト決済方法に設定するか
      responses:
        201:
          description: 決済方法追加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/PaymentMethod'
  /payment/methods/{methodId}/set-default:
    post:
      tags: [決済処理]
      summary: デフォルト決済方法設定
      parameters:
        - name: methodId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: デフォルト設定成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
  /payment/methods/{methodId}:
    delete:
      tags: [決済処理]
      summary: 決済方法削除
      parameters:
        - name: methodId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 決済方法削除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
  /payment/invoices:
    get:
      tags: [決済処理]
      summary: 請求書一覧取得
      description: |
        請求書の一覧を取得します。
        ページネーション、ステータスフィルタリング対応。
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [paid, unpaid, overdue, draft]
        - name: tenantId
          in: query
          schema:
            type: string
      responses:
        200:
          description: 請求書一覧取得成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Invoice'
  /payment/invoices/{invoiceId}:
    get:
      tags: [決済処理]
      summary: 個別請求書詳細
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 請求書詳細取得成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Invoice'
  /payment/invoices/{invoiceId}/download:
    get:
      tags: [決済処理]
      summary: 請求書PDFダウンロード
      description: |
        請求書のPDFファイルをダウンロードします。
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: PDFダウンロード成功
          content:
            application/pdf:
              schema:
                type: string
                format: binary
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token (Keycloak発行)
    
    demoMode:
      type: apiKey
      in: header
      name: X-Demo-Mode
      description: デモモード用認証 (開発・テスト環境のみ)
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          description: データ配列
        pagination:
          type: object
          properties:
            total:
              type: integer
              description: 総件数
            limit:
              type: integer
              description: 1ページあたりの件数
            offset:
              type: integer
              description: オフセット
            has_more:
              type: boolean
              description: 次ページの有無

    # =============================================================================
    # 認証関連
    # =============================================================================
...
    BillingSummary:
      type: object
      properties:
        current_plan:
          type: object
          properties:
            name:
              type: string
              example: "Professional"
            price:
              type: number
              example: 99.00
            currency:
              type: string
              example: "USD"
            billing_cycle:
              type: string
              enum: [monthly, yearly]
              example: "monthly"
        usage:
          type: object
          properties:
            api_calls:
              type: object
              properties:
                current:
                  type: integer
                  example: 15420
                limit:
                  type: integer
                  example: 50000
                percentage:
                  type: number
                  example: 30.84
            users:
              type: object
              properties:
                current:
                  type: integer
                  example: 12
                limit:
                  type: integer
                  example: 50
            storage:
              type: object
              properties:
                current:
                  type: number
                  example: 2.3
                limit:
                  type: number
                  example: 100
                unit:
                  type: string
                  example: "GB"
        next_billing_date:
          type: string
          format: date
          example: "2024-11-01"
        payment_method:
          $ref: '#/components/schemas/PaymentMethod'
    PaymentMethod:
      type: object
      properties:
        id:
          type: string
          example: "pm_1234567890"
        type:
          type: string
          enum: [card, bank_account]
          example: "card"
        card:
          type: object
          properties:
            brand:
              type: string
              example: "visa"
            last4:
              type: string
              example: "4242"
            exp_month:
              type: integer
              example: 12
            exp_year:
              type: integer
              example: 2025
        is_default:
          type: boolean
          example: true
        stripe_payment_method_id:
          type: string
        created_at:
          type: string
          format: date-time
    Invoice:
      type: object
      properties:
        id:
          type: string
          example: "inv_1234567890"
        tenant_id:
          type: string
        amount_total:
          type: number
          example: 99.00
        amount_paid:
          type: number
          example: 99.00
        currency:
          type: string
          example: "USD"
        status:
          type: string
          enum: [paid, unpaid, overdue, draft]
          example: "paid"
        stripe_invoice_id:
          type: string
        invoice_number:
          type: string
          example: "INV-2024-001"
        billing_period_start:
          type: string
          format: date
        billing_period_end:
          type: string
          format: date
        due_date:
          type: string
          format: date
        paid_at:
          type: string
          format: date-time
        invoice_pdf:
          type: string
          format: uri
        line_items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              quantity:
                type: integer
              unit_amount:
                type: number
              amount:
                type: number
        created_at:
          type: string
          format: date-time
...